duplicated classes removed & minor refactoring
